server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyautogui==0.9.54
pystray==0.19.5
pillow==10.1.0
pyinstaller==6.3.0